move expanded granular
