--- feature_ext_sc/cli/slicer_cli_list.json
@@ -1,5 +1,0 @@
-{
-    "ExpandedGranularFeatures": {
-      "type"    : "python"
-    }  
-}
