+++ fextract/cli/slicer_cli_list.json
@@ -1,5 +1,8 @@
 {
-       "ClassicalFeatures": {
+    "ClassicalFeatures": {
+      "type"    : "python"
+    },
+    "ExpandedGranularFeatures": {
       "type"    : "python"
     }
   }
